refactor: move `Birds` from root to Renderer folder and fix include paths
