+++ source/app/app_game.cpp
@@ -16,7 +16,7 @@
 #include "Garages.h"
 #include "UIRenderer.h"
 #include "Gamma.h"
-#include "Birds.h"
+#include "../game_sa/Renderer/Birds.h"
 #include "Skidmarks.h"
 #include "Ropes.h"
 #include "Glass.h"
