+++ source/game_sa/Game.cpp
@@ -7,7 +7,7 @@
 #include "TheCarGenerators.h"
 #include "Radar.h"
 #include "Credits.h"
-#include "Birds.h"
+#include "Renderer/Birds.h"
 #include "MovingThings.h"
 #include "PlantMgr.h"
 #include "Occlusion.h"
