+++ source/game_sa/Renderer/Birds.h
